requirements.txt
pydantic
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-core
langchain-community
defusedxml 
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-community
faiss-cpu
langchain_google_genai
pypdf
langchain_huggingface
langchain
langchain-pinecone
python-docx
cohere
pandas
rank_bm25
nltk
streamlit